backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
langgraph
langchain-google-genai
pydantic
python-multipart
python-dotenv
